Refs #368 - extension to expand shortened urls
This code does not work at all (yet). I created an extension which
detects shortened urls in status updates, and uses 'longurlplease' to get
the destination of these urls.

What it does not do yet, is replacing the short urls in the updates by
the long ones. I put a lot of comments in the extension code, which
describe the problems I still have to cope with.