Closes #368 - plugin for expanding short urls
Although I do not understand why, it is working for Twitter as well,
despite the strange t.co url shortener Twitter used.

I changed the behaviour of the extension so that short urls are still
shown as short urls, but when you 'hover' over a short url, the long
url pops up. When the short url is clicked, the long url is opened,
skipping the url shortener service.

I also added tinyurl.com to the regexp.