Closes #411 - expandurls and b1t.it, ow.ly
Changing the regexp was not a very big hack. But I changed the
'unshortening service' from longurlplease to longurl.org as well.
For two reasons:

* it supports b1t.it
* it is open source, so if in the feature we need support for
  yet another shortener, we can add it ourselves if needed :)

longurl.org does not provide the short url in its results, so I
had to change the service call into a direct jquery ajax call
with a context. I am not sure whether this can be done with the
functions in lib.network.js.