Fix Issue #482.
Perspectival boolean attributes in timelines like these are "best
effort", especially for embedded tweets. After some tests, I found that
the outer values seems more reliable, so I copied them to the embedded
one.

See https://gist.github.com/leafduo/5023250 for a test case.
This JSON object was captured from streaming api after @yukixz
favorited and retweeted the original tweet. The favorite field of
embedded tweet is `true`, which was _wrong_ since I didn't favorited
that. But the outer one was correct. BTW, after a short period of
time, the one in embedded tweet became `false`, which was correct.
So, this may be an issue trading consistency for scalability.

Useful links:
https://dev.twitter.com/discussions/11632
https://dev.twitter.com/discussions/14351