Fix startup hang with recent webkit
For some reason the chrome integration triggers with certain webkit
versions in the GTK (and maybe Qt) application.

The startup hangs at "Loading Resources" because a reference to "chrome"
cannot be resolved and triggers an uncaught exception locking up the
startup.

I wrapped the call to chrome in a try{} catch{} block to make it pass.
Shouldn't break any chrome integration but will get unchromy things to
startup properly.