Fix for avatar image not being saved in disk cache on GTK version
== Symptom

Avatar icon in the desktop notification (libnotify) of GTK version was missing.

== Cause

Save to disk operation was aborted due to an exception in the `get_prefs`
function on agent.py, as it was trying to fetch a non-existing element
(proxy_type) from config.settings dictionary.

== Solution

Check if the key exists in the dictrionary before using it, and return an empty
string if not.