Fix #4. Read extended media entities for media preview
A while ago, Twitter added the ability to upload multiple images to a
tweet. Hotot didn't support this and just displays the first image for a
tweet.

Data about entities are placed in the 'entities' key of the response.
However, the data there only contains information for the first image.
It turns out that Twitter added an extra key called 'extended_entities',
which contains data on all uploaded images. Thus, the bug is fixed by
modifying Hotot to read from 'extended_entities' instead of 'entities'.